refactor: organize functions into dir/files
