+++ src/components/index.ts
@@ -1,0 +1,4 @@
+export { Causality } from "./causality"
+export { Cell } from "./cell"
+export { SpaceTime } from "./space-time"
+export { SpatialSubstrate } from "./spatial-substrate"
