+++ src/constants.ts
@@ -1,0 +1,5 @@
+export const LIFE = 1
+export const DEATH = 0
+export const FIRST_DIMENSION = 23
+export const GENERATION = 80
+export const T = 0
